src/test/java/practice/leecode/LinkedTable/MiddleNode_876Test.java
package practice.leecode.LinkedTable;

import org.junit.Test;
import practice.utils.ListNode;

import static org.junit.Assert.*;

public class MiddleNode_876Test {

    @Test
    public void middleNode1() {

        ListNode node1 = new ListNode(1);
        ListNode node2 = new ListNode(2);
        ListNode node3 = new ListNode(3);
        ListNode node4 = new ListNode(4);
        ListNode node5 = new ListNode(5);

        node1.next = node2;
        node2.next = node3;
        node3.next = node4;
        node4.next = node5;

        MiddleNode_876 middleNode_876 = new MiddleNode_876();
        assertEquals(3, middleNode_876.middleNode(node1).val);

    }

    @Test
    public void middleNode2() {

        ListNode node1 = new ListNode(1);
        ListNode node2 = new ListNode(2);
        ListNode node3 = new ListNode(3);
        ListNode node4 = new ListNode(4);
        ListNode node5 = new ListNode(5);
        ListNode node6 = new ListNode(6);

        node1.next = node2;
        node2.next = node3;
        node3.next = node4;
        node4.next = node5;
        node5.next = node6;

        MiddleNode_876 middleNode_876 = new MiddleNode_876();
        assertEquals(4, middleNode_876.middleNode(node1).val);

    }
}